Phase 7: Jupiter Swap
implemented Jupiter swap when the token selected is not SOL, the players are paid in SOL.
added a readme file to correct the phase naming issue, the file address the fault in the phase naming.
updated the naming issues in all the backend files
added a test_phase7.py script to test out the swap feature.